refactor: narrow vault settlement API surface
