+++ docs/source/api/vault/index.rst
@@ -30,7 +30,6 @@ A generic high-level Python framework to integrate different vault providers.
    eth_defi.vault.curator
    eth_defi.vault.curator_export
    eth_defi.vault.settlement_data
-   eth_defi.vault.settlement_event_reader
    eth_defi.vault.settlement_scan
    eth_defi.vault.data_file_export
    eth_defi.vault.top_vaults_json
